aias/requirements.txt
requests
pyautogui
pytesseract
pygetwindow
pillow
pyyaml
